requirements.txt
django
python-dotenv
mysqlclient
djangorestframework
beautifulsoup4
bcrypt
django-cors-headers
djangorestframework-simplejwt
pyjwt
pillow
gunicorn
python-magic
imagehash
stripe
drf-spectacular